Improvements to echo command. 1. A newline controlling switch was provided. 2. Single write system call for special kernel files interaction.
Documentatation:
Usage: echo [-n] [text...] Prints the
input text to the stdout using the write system call. The command will
insert a new line at the end of the text except if the -n switch is
provided. Notes: It is important for text to be written in one write
call to allow echo to interact with special kernel files such as for
cgroups.

Change-Id: Ib968de9adb722ebc180e4008d6d6c6f2c451d93a