fixed Makefile to not use SSE opcodes
xv6 doesn't track the state of xmm registers when doing context switches. Thus we must not use them.
The bug was found when trying to compile on centos 7 and it used opcode with xmm. An illegal opcode interrupt was rasied.
To solve it, I added `-mno-sse` to the compile flags.

In addition, I redirected stderr to /dev/null while searching for qemu binary file.

Change-Id: Iad1430f288a9a10e45ce0ecc08ddfa1f6d699054