Adding xv6-objfs to xv6-ns: implement the lower level (disk driver level)
Imlpement the disk driver level with all the relevant tests.
In addition, a simple tests library was written for the tests.
Target to run the tests was add to the makefile as well as documentation in the README

In the current implementation, the driver layer is NOT thread safe.
Would be added in the future.

Change-Id: I915497d88069ad9736c299a5eb8eff2fcb1bd092