Fix crash when mounting an external fs, and add test
There was an improper loop device handling in getsuperblock function.
as a result, after the mount the OS became unstable.
In addition, added panics on errors instead returning 0.
The bug occurs only when the mount is performed from a child process.
(It occurred in the shell but not in the exist tests).
The updated tests mimics the shell:
It preforms the mount in a child process and afterwards does the necessary validations.

Change-Id: Icdfe5ae17dc44f4ddc63853876e2ef6036347f7e