bug: parent from mountpoint is mountpoint itself
There is a bug when mounting an internal fs to a dir, for example "a". The bug is that the parent from "a"'s perspective is itself instead of its actual parent.
This is reflected when changing dir into a mounted dir and then trying to cd to .., the cwd will stay the mounted dir.

In this commit I added a test to mounttest which catches the bug - cdinthenouttest.

Change-Id: Ibcb6c60bb998e9158c97728f32de5f612e280784