Fixing crashes and corruption on mounting loop dev
In the past, mounting a loop device resulted in kernel panics and fs corruption
(either instantly or upon access to the mounted dir) because the device index
was calculated incorectly. This patch fixes this behavior and enhances the
mount tests to check for faulty mounting. In addition, this patch fixes the
mount command to work for arbitrary device names.

Change-Id: Icb1f91454bbb11d67d947cbae8202312e9fcd8b9