Fixing '..' behavior for mounts
The '..' directory in the root inode of a filesystem should refer to itself,
but when the fs is mounted, the '..' directory should be traversed and point
to the parent of the mount directory. This patch implements this behavior.
In addition it contains the device index fix of mmn11 so new tests won't fail
on old bugs.

Closes-Bug: #514367
Change-Id: I21cc8b3036af477b2f0ab4b4dfe9db96f6e5788e