Added cgroup.max.descendants file - resolved stat error.
cgroup.max.descendants file was removed due to a stat not working.
The problem was that cg_open couldn't find which cg file the file was because the file name had a typo. The file name was descandants insted of descendants.
We also added a test in cgroupstests to test that all the files in a cgroup dir are indeed with a cg file name. This test will catch errors of this type.