Fix crashes in cgroupstests (maman 12)
In the past, cgroupstests would sometimes fail in various stages with
memory related errors. This patchset fixes two bugs that caused the
crashes:

 - The cgroups memory usage counters are now updated during exec(),
   preventing fork() fails due to incorrect limits;

 - A stack overflow in unsafe_cg_read() caused a memory corruption, so
   the offending buffer is now moved to a static location.

In addition, a test function for free memory is added - kmemtest(). It
can detect incorrect memory usage, such as the second bug.

This function was added as a system call - kmemtest().