Fix busy-waiting on inactive terminals
In the past, when more than one process interacted with the console
(e.g. the login shell and a pouch), the active one would sleep waiting
for input as expected, but the *inactive* one would constantly try and
fail to interact with the console, generating much overhead.

This patch fixes the issue by introducing a per-device lock for the
console and ttys, to allow sleeping until the device owns the console
(DEV_CONNECTED flag is on).

Change-Id: I250cafb53a57a7791c4b5dc42f400dc484aefd47