Add IO statistics support for cgroups
Added io.stat interface file for cgroups which gives info of what IO
devices attached to the current cgroup.
Lines are keyed by $MAJ:$MIN device numbers and not ordered.
	The following nested keys are defined.

	  ======	===================
	  rbytes	Bytes read
	  wbytes	Bytes written
	  rios		Number of read IOs
	  wios		Number of write IOs
	  ======	===================
An example of the output when reading io.stat file:
8:16 rbytes=1459200 wbytes=314773504 rios=192 wios=353

In addition to the major change a header circular dependency have beed
resolved in file.h and cgroup.h.

Note #1: For now we have only TTY IO devices which are connected to the
console (a hardware interface). For future IO devices this might come in
handy.
Note #2: The interface file is not read-only although it should be.

Change-Id: I80551695a4b67160d04dc02bdcb8e61b909c5b33