Add IO statistics support for cgroups
Added io.stat interface file for cgroups which gives info of what IO
devices attached to the current cgroup.
Lines are keyed by $MAJ:$MIN device numbers and not ordered.
	The following nested keys are defined.

	  ======	===================
	  rbytes	Bytes read
	  wbytes	Bytes written
	  rios		Number of read IOs
	  wios		Number of write IOs
	  ======	===================
An example of the output when reading io.stat file:
8:16 rbytes=1459200 wbytes=314773504 rios=192 wios=353

In addition to the major change a header circular dependency have beed
resolved in file.h and cgroup.h.

Note #1: For now we have only TTY IO devices which are connected to the
console (a hardware interface). For future IO devices this might come in
handy.
Note #2: The interface file is not read-only although it should be.

Note #3: Correct minor bugs in io.stat feature

   - Removed all the inode references in the cgroup IO functions. We don't
     use inodes now to get minor/major numbers now.

   - Renamed dev_state structure to dev_stat to indicate that it stores
     statistics.

   - Added test for io.stat feature in expect test script

   - Fixed minor issues which resulted in bugs in corner cases (for
     example, when destroying a cgroup)

Change-Id: Iba88fce5fcd96e1ea200550440a3020c35d34eaf