Add mutex support for user-space
Some user space modules and applications (like pouch) need a way for ipc
and synchronization.
For this purposes mutexes were introduced (similar to already existing
spinlocks for the kernel). The mutex is available for the user through
syscall API (init, lock, unlock and delete).

Note: for now, the user is responsible to use mutexes correctly and to
not alter the mutex structure it receives (as a black box).

Change-Id: I12d08a9470622ce9baddb6e2b86f7e643d0095cb